+++ ivy/data_classes/array/gradients.py
@@ -107,16 +107,6 @@ class _ArrayWithGradients(abc.ABC):
             self, mw, vw, step, beta1=beta1, beta2=beta2, epsilon=epsilon, out=out
         )
 
-    def adagrad_step(
-        self: ivy.Array,
-        vt: Union[ivy.Array, ivy.NativeArray],
-        /,
-        *,
-        epsilon: float = 0,
-        out: Optional[ivy.Array] = None,
-    ) -> ivy.Array:
-        return ivy.adagrad_step(self, vt, epsilon=epsilon, out=out)
-
     def optimizer_update(
         self: ivy.Array,
         effective_grad: Union[ivy.Array, ivy.NativeArray],
@@ -351,31 +341,6 @@ class _ArrayWithGradients(abc.ABC):
             beta1=beta1,
             beta2=beta2,
             epsilon=epsilon,
-            stop_gradients=stop_gradients,
-            out=out,
-        )
-
-    def adagrad_update(
-        self: ivy.Array,
-        dcdw: Union[ivy.Array, ivy.NativeArray],
-        lr: Union[float, ivy.Array, ivy.NativeArray],
-        vt_tm1: Union[ivy.Array, ivy.NativeArray],
-        step: int,
-        /,
-        *,
-        epsilon: float = 1e-7,
-        lr_decay: float = 0,
-        stop_gradients: bool = True,
-        out: Optional[ivy.Array] = None,
-    ) -> ivy.Array:
-        return ivy.adagrad_update(
-            self,
-            dcdw,
-            lr,
-            vt_tm1,
-            step,
-            epsilon=epsilon,
-            lr_decay=lr_decay,
             stop_gradients=stop_gradients,
             out=out,
         )
