+++ ivy_tests/test_ivy/test_functional/test_experimental/test_core/test_gradients.py
@@ -1,10 +1,14 @@
 # global
+from hypothesis import strategies as st
 import pytest
 import numpy as np
 
 # local
+from ivy_tests.test_ivy.test_functional.test_core.test_gradients import (
+    get_gradient_arguments_with_lr,
+)
 import ivy_tests.test_ivy.helpers as helpers
-from ivy_tests.test_ivy.helpers.pipeline_helper import update_backend
+from ivy_tests.test_ivy.helpers import handle_test, update_backend
 
 
 # bind_custom_gradient_function
@@ -49,3 +53,84 @@ def test_bind_custom_gradient_function(
     for grad, grad_from_gt in zip(grad_np, grad_np_from_gt):
         assert grad.shape == grad_from_gt.shape
         assert np.allclose(grad, grad_from_gt)
+
+
+@handle_test(
+    fn_tree="functional.ivy.experimental.adagrad_step",
+    dtype_n_dcdw_n_vt=get_gradient_arguments_with_lr(
+        num_arrays=2,
+        no_lr=True,
+        min_value=1e-05,
+        max_value=1e08,
+        large_abs_safety_factor=2,
+        small_abs_safety_factor=2,
+    ),
+    epsilon=st.floats(min_value=1e-1, max_value=1),
+)
+def test_adagrad_step(
+    *,
+    dtype_n_dcdw_n_vt,
+    epsilon,
+    test_flags,
+    backend_fw,
+    fn_name,
+    on_device,
+):
+    input_dtypes, [dcdw, vt] = dtype_n_dcdw_n_vt
+    helpers.test_function(
+        input_dtypes=input_dtypes,
+        test_flags=test_flags,
+        backend_to_test=backend_fw,
+        fn_name=fn_name,
+        on_device=on_device,
+        rtol_=1e-1,
+        atol_=1e-1,
+        dcdw=dcdw,
+        vt=vt,
+        epsilon=epsilon,
+    )
+
+
+# adagrad_update
+@handle_test(
+    fn_tree="functional.ivy.experimental.adagrad_update",
+    dtype_n_ws_n_dcdw_n_vt_n_lr=get_gradient_arguments_with_lr(
+        num_arrays=3,
+        min_value=1e-05,
+        max_value=1e08,
+        large_abs_safety_factor=2.0,
+        small_abs_safety_factor=2.0,
+    ),
+    step=st.integers(min_value=1, max_value=10),
+    epsilon=st.floats(min_value=1e-2, max_value=1),
+    stopgrad=st.booleans(),
+)
+def test_adagrad_update(
+    *,
+    dtype_n_ws_n_dcdw_n_vt_n_lr,
+    step,
+    epsilon,
+    stopgrad,
+    test_flags,
+    backend_fw,
+    fn_name,
+    on_device,
+):
+    input_dtypes, [w, dcdw, vt], lr = dtype_n_ws_n_dcdw_n_vt_n_lr
+    stop_gradients = stopgrad
+    helpers.test_function(
+        input_dtypes=input_dtypes,
+        test_flags=test_flags,
+        backend_to_test=backend_fw,
+        fn_name=fn_name,
+        on_device=on_device,
+        rtol_=1e-2,
+        atol_=1e-2,
+        w=w,
+        dcdw=dcdw,
+        lr=lr,
+        vt_tm1=vt,
+        step=step,
+        epsilon=epsilon,
+        stop_gradients=stop_gradients,
+    )
